Refactor recent merge with changes to Text (from Native)
