+++ packages/components/src/text/styles/emotion-css.js
@@ -1,6 +1,6 @@
 /**
  * External dependencies
  */
-import css from '@emotion/css';
+import { css } from '@emotion/core';
 
 export default css;
