+++ packages/components/src/text/styles/font-family.js
@@ -1,0 +1,2 @@
+export const fontFamily = `font-family: -apple-system, BlinkMacSystemFont, "Segoe UI", Roboto,
+Oxygen-Sans, Ubuntu, Cantarell, "Helvetica Neue", sans-serif;`;
