+++ packages/components/src/text/styles/font-family.native.js
@@ -1,0 +1,1 @@
+export const fontFamily = '';
